feat(zed): Manage zed config files with nix
All we're doing at the moment is managing the dotfiles with Nix, Zed is
so actively maintained at the moment that it's easier to just let it
keep itself up-to-date.

Once there are stable releases it might make a bit more sense to manage
the binary with Nix as well.